refact: split i18n code to WebUIi18nSettings
